created new logo, created populate route for stocks and populated db
